requirements.txt
selenium~=4.11.2
allure-pytest~=2.13.1
requests~=2.28.2
